Contacts complete

+++ src/services/v1/contact/removeContact.js
@@ -1,8 +1,26 @@
+/**
+ * @description Deletes a Contact
+ *
+ * @module removeContact
+ *
+ * @requires Contact
+ * @requires createQueryResponse
+ * @requires searchContacts
+ *
+ * @param {string} id - Contact id
+ *
+ * @returns {responseTemplate} Custom JSON status response
+ */
 const Contact = require("../../../db/models/v1/Contact");
 const createQueryResponse = require("../createQueryResponse");
+const searchContacts = require("./searchContacts");
 
 const removeContact = async (id) => {
-  // let result = {};
+  let checkResult = await searchContacts.byId(id);
+  console.log(checkResult);
+  if (checkResult.status === 404) {
+    return createQueryResponse.doesNotExist(true, checkResult);
+  }
 
   try {
     const result = await Contact.findByIdAndDelete(id);
@@ -10,7 +28,6 @@ const removeContact = async (id) => {
   } catch (err) {
     return createQueryResponse.deleteResponse(true, null, err);
   }
-  // return result;
 };
 
 module.exports = removeContact;
